requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
solana==0.32.0
solders==0.20.0
aiohttp==3.9.1
asyncio==3.4.3
python-dotenv==1.0.0
pydantic==2.5.2
base58==2.1.1
cryptography==41.0.8
requests==2.31.0
websockets==12.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
